refactor: split SwiftUIRenderingCoverageTests to silence type_body_length warning
- Extract library, scores, server songs, and session results tests into
  SwiftUIRenderingLibraryAndResultsTests.swift
- Retain settings, audio, notation primitives, and gameplay layout in
  SwiftUIRenderingCoverageTests.swift
- Extract settings, input, metronome, and gameplay header tests into
  SwiftUIRenderingSettingsAndHeaderTests.swift (from prior commit)
- All three files now pass SwiftLint with zero violations
